% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_country_arg}
\alias{check_country_arg}
\title{Check the `country` argument used in load_ots() and load_rts().}
\usage{
check_country_arg(country)
}
\arguments{
\item{country}{The `country` argument.}
}
\description{
Check the `country` argument used in load_ots() and load_rts().
}
