% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_flow_arg}
\alias{check_flow_arg}
\title{Check the `flow` argument used in load_ots() and load_rts().}
\usage{
check_flow_arg(flow, valid = 1:4)
}
\arguments{
\item{flow}{The `flow` argument.}

\item{valid}{The set of flow codes actually valid for the calling
function. Defaults to 1:4 (all four OTS flow codes). `load_rts()`
passes `3:4`, since RTS data is only ever populated for those two
codes - see its documentation for details.}
}
\description{
Check the `flow` argument used in load_ots() and load_rts().
}
