% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_internet.R
\name{check_internet}
\alias{check_internet}
\title{A function to check for internet connection. Used within `load_*()` functions.}
\usage{
check_internet()
}
\description{
A function to check for internet connection. Used within `load_*()` functions.
}
