% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_logical_arg}
\alias{check_logical_arg}
\title{Check a single TRUE/FALSE argument.}
\usage{
check_logical_arg(value, name)
}
\arguments{
\item{value}{The value to check. NULL is allowed (treated as "not set").}

\item{name}{The argument name, used in the error message.}
}
\description{
Check a single TRUE/FALSE argument.
}
