% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_month_arg}
\alias{check_month_arg}
\title{Check the `month` argument used in load_ots() and load_rts().}
\usage{
check_month_arg(month)
}
\arguments{
\item{month}{The `month` argument.}
}
\description{
Check the `month` argument used in load_ots() and load_rts().
}
