% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_output_arg}
\alias{check_output_arg}
\title{Check the `output` argument used across all load_*() functions.}
\usage{
check_output_arg(output)
}
\arguments{
\item{output}{The `output` argument.}
}
\description{
Check the `output` argument used across all load_*() functions.
}
