% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_region_arg}
\alias{check_region_arg}
\title{Check the `region` argument used in load_ots() and load_rts().}
\usage{
check_region_arg(region)
}
\arguments{
\item{region}{The `region` argument.}
}
\description{
Check the `region` argument used in load_ots() and load_rts().
}
