% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_status.R
\name{check_status}
\alias{check_status}
\title{A function to check the API's response. Used within `load_*()` functions.}
\usage{
check_status(resp)
}
\arguments{
\item{resp}{The API response obtained from https://api.uktradeinfo.com, as
returned by `httr2::req_perform()`.}
}
\description{
A function to check the API's response. Used within `load_*()` functions.
}
