% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_suppression_arg}
\alias{check_suppression_arg}
\title{Check the `suppression` argument used in load_ots().}
\usage{
check_suppression_arg(suppression)
}
\arguments{
\item{suppression}{The `suppression` argument.}
}
\description{
Check the `suppression` argument used in load_ots().
}
