% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_uk_country_arg}
\alias{check_uk_country_arg}
\title{Check the `uk_country` argument used in load_rts().}
\usage{
check_uk_country_arg(uk_country)
}
\arguments{
\item{uk_country}{The `uk_country` argument.}
}
\description{
Check the `uk_country` argument used in load_rts().
}
