% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_filter.R
\name{create_filter}
\alias{create_filter}
\title{A function to create a filter from function arguments. Used within `load_ots()` and `load_rts` functions.}
\usage{
create_filter(list)
}
\arguments{
\item{list}{A list of arguments obtained within the `load_rts()` or `load_ots()` functions.}
}
\description{
A function to create a filter from function arguments. Used within `load_ots()` and `load_rts` functions.
}
