% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_status.R
\name{check_status}
\alias{check_status}
\title{A function to check the API's response. Used within `load_*()` functions.}
\usage{
check_status(res)
}
\arguments{
\item{res}{The API response obtained from https://api.uktradeinfo.com.}
}
\description{
A function to check the API's response. Used within `load_*()` functions.
}
