% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ots.R
\name{load_ots}
\alias{load_ots}
\title{A function for loading OTS data via HMRC's API.}
\usage{
load_ots(
  month = NULL,
  flow = NULL,
  commodity = NULL,
  sitc = NULL,
  country = NULL,
  region = NULL,
  uk_port = NULL,
  suppression = NULL,
  join_lookup = TRUE,
  print_url = FALSE,
  output = "tibble",
  skip_interval = 40000,
  use_proxy = FALSE,
  ...
)
}
\arguments{
\item{month}{The month(s) to be loaded in the form of a vector of two integers (YYYYMM), where the first element is the minimum date, and second the maximum date. Defaults to NULL (all months).}

\item{flow}{The trade flow to be loaded. Takes one ore more integers (1, 2, 3, and/or 4), where 1 is EU imports, 2 is EU exports, 3 is non-EU imports, and 4 is non-EU exports. Defaults to NULL (all flows).}

\item{commodity}{One or more HS2, HS4, HS6, or CN8 commodity codes in the form of a numeric or character vector. Defaults to NULL (all commodities). Any missing leading zeros will be automatically added.}

\item{sitc}{One or more SITC1, SITC2, SITC3, SITC4, or SITC5 commodity codes in the form of a character vector. Defaults to NULL (all commodities).}

\item{country}{One or more destination or origin countries by their 2-letter ISO code. Defaults to NULL (all countries).}

\item{region}{One or more destination or origin regions. Defaults to NULL (all regions). Takes one or more of the following broad categories: "Asia and Oceania", "Eastern Europe exc EU", "European Union", "Latin America and Caribbean", "Middle East and N Africa", "North America", "Sub-Saharan Africa", "Western Europe exc EU", "Western Europe exc EC", "Low Value Trade", "Stores and Provisions", and/or "Confidential Region".}

\item{uk_port}{One or more departure or arrival ports by their three-letter code (only available for trade with non-EU countries prior to 2021, and all trade post-2021). Defaults to NULL (all ports in the UK).}

\item{suppression}{One or more suppression codes. Takes one or more integers between 1 and 5 (see HMRC API guidance for information). Defaults to NULL (all available results).}

\item{join_lookup}{A logical value indicating whether results should be joined with lookups from the API. Defaults to TRUE. Setting to FALSE will return a smaller but less human-readable dataframe containing only codes.}

\item{print_url}{A logical. Defaults to FALSE. Setting this to TRUE will print the URL(s) used to load the trade data to the console.}

\item{output}{A character specifying if a tibble ("tibble") or dataframe ("df") should be returned. Defaults to "tibble".}

\item{skip_interval}{Passed to load_custom(). A non-negative integer value showing the skip interval for paginated results. Defaults to 40,000 rows.}

\item{use_proxy}{A logical. Defaults to FALSE. Setting this to TRUE will allow the use of a proxy connection using `use_proxy()` from `httr`.}

\item{...}{Optional arguments to be passed along to `use_proxy()` when using a proxy connection (by setting use_proxy to TRUE). See the `httr` documentation for more details.}
}
\value{
Returns a dataframe or tibble
}
\description{
A function for loading OTS data via HMRC's API.
}
\examples{
\dontrun{
# Obtaining all trade of single malt Scotch whisky and bottled gin between in
# 2019 via the OTS endpoint:

load_ots(month = c(201901, 201912), commodity = c(22083030, 22085011))

}
}
\keyword{api}
\keyword{data}
\keyword{hmrc}
\keyword{overseas}
\keyword{statistics}
\keyword{trade}
