% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rts.R
\name{load_rts}
\alias{load_rts}
\title{A function for loading RTS data via HMRC's API.}
\usage{
load_rts(
  month = NULL,
  flow = c(1, 2, 3, 4),
  sitc = NULL,
  country = NULL,
  region = NULL,
  uk_country = NULL,
  join_lookup = TRUE,
  output = "tibble"
)
}
\arguments{
\item{month}{The month(s) to be loaded in the form of a vector of two integers (YYYYMM), where the first element is the minimum date, and second the maximum date. Defaults to NULL (all months).}

\item{flow}{The trade flow to be loaded. Takes one ore more integers (1, 2, 3, and/or 4), where 1 is EU imports, 2 is EU exports, 3 is non-EU imports, and 4 is non-EU exports. Defaults to NULL (all flows).}

\item{sitc}{The range of SITC2 commodity codes to be loaded (min and max), in the form of a numeric vector. Defaults to NULL (all commodities).}

\item{country}{One or more destination or origin countries by their 2-letter ISO code. Defaults to NULL (all countries).}

\item{region}{One or more destination or origin regions. Defaults to NULL (all regions). Takes one or more of the following broad categories: "Asia and Oceania", "Eastern Europe exc EU", "European Union", "Latin America and Caribbean", "Middle East and N Africa", "North America", "Sub-Saharan Africa", "Western Europe exc EU", "Western Europe exc EC", "Low Value Trade", "Stores and Provisions", and/or "Confidential Region".}

\item{uk_country}{One or more destination or origin UK countries. Defaults to NULL (all countries). Takes one or more of the following: "England", "Wales", "Scotland", "Northern Ireland", and/or "Unallocated". England may have multiple regions within it, and Unallocated may be split between known and unknown.}

\item{join_lookup}{A logical value indicating whether results should be joined with lookups from the API. Defaults to TRUE. Setting to FALSE will return a smaller but less human-readable dataframe containing only codes.}

\item{output}{A character specifying if a tibble ("tibble") or dataframe ("df") should be returned. Defaults to "tibble".}
}
\value{
Returns a dataframe or tibble
}
\description{
A function for loading RTS data via HMRC's API.
}
\examples{
\dontrun{
# Obtaining all trade from SITC2 code 00 - Live animals to
# 11 - Beverages in 2019 via the RTS endpoint:

load_rts(month = c(201901, 201912), sitc = c(00, 11))

}
}
\keyword{api}
\keyword{data}
\keyword{hmrc}
\keyword{regional}
\keyword{statistics}
\keyword{trade}
